requirements.txt
astroid==2.12.13
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
autopep8==2.0.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
channels==3.0.1
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
daphne==3.0.2
Deprecated==1.2.13
dill==0.3.6
Django==3.2.15
django-cors-headers==3.10.1
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
flake8==5.0.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
msgpack==1.0.4
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==1.7.1
pylint==2.15.6
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stripe==5.0.0
tomli==2.0.1
tomlkit==0.11.6
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.12
wget==3.2
